Lots of fixes (constants, moderation, yml, etc)

+++ src/main/java/blogengine/models/dto/blogdto/tagdto/SingleTagDto.java
@@ -8,5 +8,5 @@ import lombok.Data;
 public class SingleTagDto {
 
     private String name;
-    private double weight;
+    private float weight;
 }
